[fix] entitiy 필드 타입 수정

+++ src/main/java/com/example/myboard/domain/board/dto/response/BoardCommentResponse.java
@@ -7,7 +7,7 @@ public record BoardCommentResponse(
         String commentMemberEmail,
         String ParentMemberEmail,
         String comment,
-        Integer boardId
+        Long boardId
 ) {
 
     public static BoardCommentResponse fromEntity(BoardComment boardComment){
